requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Session==0.4.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
mypy==1.2.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
openpyxl==3.1.2
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
ruff==0.0.269
six==1.16.0
SQLAlchemy==2.0.12
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.3
